`remove_extra_metadata` etc `transforms/utils.py` -> `data/utils.py`
