+++ monai/data/__init__.py
@@ -66,6 +66,7 @@ from .utils import (
     create_file_basename,
     decollate_batch,
     dense_patch_slices,
+    get_extra_metadata_keys,
     get_random_patch,
     get_valid_patch_size,
     is_supported_format,
@@ -79,6 +80,8 @@ from .utils import (
     partition_dataset_classes,
     pickle_hashing,
     rectify_header_sform_qform,
+    remove_extra_metadata,
+    remove_keys,
     reorient_spatial_axes,
     resample_datalist,
     select_cross_validation_folds,
