+++ monai/transforms/__init__.py
@@ -569,7 +569,6 @@ from .utils import (
     generate_label_classes_crop_centers,
     generate_pos_neg_label_crop_centers,
     generate_spatial_bounding_box,
-    get_extra_metadata_keys,
     get_extreme_points,
     get_largest_connected_component_mask,
     get_number_image_type_conversions,
@@ -583,8 +582,6 @@ from .utils import (
     map_spatial_axes,
     print_transform_backends,
     rand_choice,
-    remove_extra_metadata,
-    remove_keys,
     rescale_array,
     rescale_array_int_max,
     rescale_instance_array,
