+++ monai/transforms/utils.py
@@ -105,9 +105,6 @@ __all__ = [
     "convert_pad_mode",
     "convert_to_contiguous",
     "get_unique_labels",
-    "remove_keys",
-    "remove_extra_metadata",
-    "get_extra_metadata_keys",
 ]
 
 
@@ -1574,69 +1571,6 @@ def convert_to_contiguous(data, **kwargs):
     if isinstance(data, Sequence):
         return [convert_to_contiguous(i, **kwargs) for i in data]
     return data
-
-
-def remove_keys(data: dict, keys: List[str]) -> None:
-    """
-    Remove keys from a dictionary. Operates in-place so nothing is returned.
-
-    Args:
-        data: dictionary to be modified.
-        keys: keys to be deleted from dictionary.
-
-    Returns:
-        `None`
-    """
-    for k in keys:
-        _ = data.pop(k, None)
-
-
-def remove_extra_metadata(meta: dict) -> None:
-    """
-    Remove extra metadata from the dictionary. Operates in-place so nothing is returned.
-
-    Args:
-        meta: dictionary containing metadata to be modified.
-
-    Returns:
-        `None`
-    """
-    keys = get_extra_metadata_keys()
-    remove_keys(data=meta, keys=keys)
-
-
-def get_extra_metadata_keys() -> List[str]:
-    """
-    Get a list of unnecessary keys for metadata that can be removed.
-
-    Returns:
-        List of keys to be removed.
-    """
-    keys = [
-        "srow_x",
-        "srow_y",
-        "srow_z",
-        "quatern_b",
-        "quatern_c",
-        "quatern_d",
-        "qoffset_x",
-        "qoffset_y",
-        "qoffset_z",
-        "dim",
-        "pixdim",
-        *[f"dim[{i}]" for i in range(8)],
-        *[f"pixdim[{i}]" for i in range(8)],
-    ]
-
-    # TODO: it would be good to remove these, but they are currently being used in the
-    # codebase.
-    # keys += [
-    #     "original_affine",
-    #     "spatial_shape",
-    #     "spacing",
-    # ]
-
-    return keys
 
 
 if __name__ == "__main__":
